Working prototype of generic manual entry workbook
Finished generic implementation of creating an Excel workbook from a collection of data model types. This Excel workbook is meant to provide contributors with a simple method for completing the raw dataset for a subsite.
