+++ Utilities.Games.ImportTool/Views/Subsites/SubsiteView.cs
@@ -25,26 +25,37 @@ namespace Utilities.Games.ImportTool.Views.Subsites
         [DynamicViewOption(nameof(_saveDataModelWorkbook_message), nameof(_saveDataModelWorkbook_color))]
         public virtual void SaveDataModelWorkbook() {
             Consoul.Write("Constructing package...");
-            var package = new ExcelPackage();
-            foreach (var datasetType in Source.DatasetTypes)
-            {
-                Consoul.Write($"\t{datasetType.Name}...", writeLine: false);
-                var typeSheet = new TypeSheet(datasetType);
-                typeSheet.BuildWorksheet(package);
-                Consoul.Write("Done!", ConsoleColor.Green);
-            }
             var fi = new FileInfo($"Utilities-Games Data Model Workbook.xlsx");
-            Consoul.Write("Saving...", writeLine: false);
-            package.SaveAs(fi);
+            using (var package = new ExcelPackage())
+            {
+                foreach (var datasetType in Source.DatasetTypes)
+                {
+                    Consoul.Write($"\t{datasetType.Name}...", writeLine: false);
+                    var typeSheet = new TypeSheetUtility(datasetType);
+                    typeSheet.BuildWorksheet(package);
+                    Consoul.Write("Done!", ConsoleColor.Green);
+                }
+                Consoul.Write("Saving...", writeLine: false);
+                package.SaveAs(fi);
+            }
             Consoul.Write("Done!", ConsoleColor.Green);
             Consoul.Write("Saved to\r\n\t" + fi.FullName, ConsoleColor.Green);
             Consoul.Wait();
 
         }
 
+        /// <summary>
+        /// Generic definition of a new Consoul view for a Subsite
+        /// </summary>
         public interface ISubsiteViewModel {
+            /// <summary>
+            /// Reference to the view title/banner.
+            /// </summary>
             string Title { get; set; }
 
+            /// <summary>
+            /// Collection of data model types that can be processed into an Excel workbook for easy manual entry.
+            /// </summary>
             IEnumerable<Type> DatasetTypes { get; set; }
         }
     }
