+++ Utilities.Games.Models/Subsites/LOTR_RiseToWar/Faction.cs
@@ -1,4 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using Utilities.Games.Models.Contracts.Attributes;
 
 namespace Utilities.Games.Models.Subsites.LOTR_RiseToWar
 {
@@ -16,6 +17,7 @@ namespace Utilities.Games.Models.Subsites.LOTR_RiseToWar
         /// <summary>
         /// Name of the first, default commander for the Faction.
         /// </summary>
+        [PseudoForeignKey(typeof(Commander), nameof(Commander.Name))]
         public string FirstCommander { get; set; }
     }
 }
