+++ Utilities.Games.Models/Subsites/LOTR_RiseToWar/RaceType.cs
@@ -5,7 +5,7 @@ namespace Utilities.Games.Models.Subsites.LOTR_RiseToWar
     /// <summary>
     /// Race of a commander.
     /// </summary>
-    public class Race
+    public class RaceType
     {
         /// <summary>
         /// Name of the race.
